PX_UNITの解釈を今頃正しく。簡易フォグ実装
git-svn-id: file:///home/gecchi/repos@2047 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/src/jp/ggaf/dx9core/util/GgafDx9GeometryScaler.cpp
@@ -55,14 +55,14 @@ void GgafDx9GeometryScaler::behave() {
             _scale[axis] += _velo_scale[axis];
             if (_top_scale[axis] <= _scale[axis]) {
                 _scale[axis] = _top_scale[axis];
-                _velo_scale[axis] = -2 * (int)(_top_scale[axis] - _bottom_scale[axis]) / (int)(_beat_spend_frame[axis]);
+                _velo_scale[axis] = -2 * (_top_scale[axis] - _bottom_scale[axis]) / (int)_beat_spend_frame[axis];
                 _one_way_cnt[axis]++;
                 if (_one_way_cnt[axis] == _stop_one_way_num[axis]) {
                     _method[axis] = NOSCALE;
                 }
             } else if (_bottom_scale[axis] >= _scale[axis]) {
                 _scale[axis] = _bottom_scale[axis];
-                _velo_scale[axis] = 2 * (int)(_top_scale[axis] - _bottom_scale[axis]) / (int)(_beat_spend_frame[axis]);
+                _velo_scale[axis] = 2 * (_top_scale[axis] - _bottom_scale[axis]) / (int)_beat_spend_frame[axis];
                 _one_way_cnt[axis]++;
                 if (_one_way_cnt[axis] == _stop_one_way_num[axis]) {
                     _method[axis] = NOSCALE;
@@ -73,7 +73,7 @@ void GgafDx9GeometryScaler::behave() {
             _scale[axis] += _velo_scale[axis];
             if (_beat_begin_frame[axis] + _beat_attack_frame[axis] == _pActor->_frame_of_active) { //アタック頂点時
                 _scale[axis] = _top_scale[axis];
-                _velo_scale[axis] = ((int)(_bottom_scale[axis] - _top_scale[axis])) / ((int)(_beat_spend_frame[axis] - _beat_attack_frame[axis] - _beat_rest_frame[axis]));
+                _velo_scale[axis] = (_bottom_scale[axis] - _top_scale[axis]) / ((int)_beat_spend_frame[axis] - (int)_beat_attack_frame[axis] - (int)_beat_rest_frame[axis]);
                 _one_way_cnt[axis]++;
                 if (_one_way_cnt[axis] == _stop_one_way_num[axis]) {
                     _method[axis] = NOSCALE;
@@ -83,7 +83,7 @@ void GgafDx9GeometryScaler::behave() {
                 _velo_scale[axis] = 0;
             } else if (_beat_begin_frame[axis] + _beat_spend_frame[axis] == _pActor->_frame_of_active) { //ループ終了時
                 _beat_begin_frame[axis] = _pActor->_frame_of_active;
-                _velo_scale[axis] = ((int)(_top_scale[axis] - _bottom_scale[axis])) / (int)(_beat_attack_frame[axis]);
+                _velo_scale[axis] = (_top_scale[axis] - _bottom_scale[axis]) / (int)_beat_attack_frame[axis];
                 _one_way_cnt[axis]++;
                 if (_one_way_cnt[axis] == _stop_one_way_num[axis]) {
                     _method[axis] = NOSCALE;
@@ -165,7 +165,7 @@ void GgafDx9GeometryScaler::loopLiner(int prm_axis, DWORD prm_beat_spend_frame, 
     _one_way_cnt[prm_axis] = 0;
     _stop_one_way_num[prm_axis] = (int)(prm_beat_num*2.0f);
     _beat_spend_frame[prm_axis] = prm_beat_spend_frame;
-    _velo_scale[prm_axis] = (int)(_top_scale[prm_axis] - _scale[prm_axis]) / ((int)prm_beat_spend_frame / 2);
+    _velo_scale[prm_axis] = (_top_scale[prm_axis] - _scale[prm_axis]) / ((int)prm_beat_spend_frame / 2);
     if (_velo_scale[prm_axis] == 0) {
         _velo_scale[prm_axis] = _top_scale[prm_axis] - _scale[prm_axis];
     }
@@ -188,7 +188,7 @@ void GgafDx9GeometryScaler::beat(int prm_axis, DWORD prm_beat_spend_frame, DWORD
     _beat_begin_frame[prm_axis] = _pActor->_frame_of_active;
     _beat_spend_frame[prm_axis] = prm_beat_spend_frame;
 
-    _velo_scale[prm_axis] = (int)(_top_scale[prm_axis] - _scale[prm_axis]) / (int)prm_attack_frame;
+    _velo_scale[prm_axis] = (_top_scale[prm_axis] - _scale[prm_axis]) / (int)prm_attack_frame;
     if (_velo_scale[prm_axis] == 0) {
         _velo_scale[prm_axis] = _top_scale[prm_axis] - _scale[prm_axis];
     }
